refactor: 외부 요청하는 infra영역 공통으로 수정
