+++ src/main/java/kitchenpos/common/infra/DefaultPurgomalumClient.java
@@ -1,5 +1,6 @@
-package kitchenpos.infra;
+package kitchenpos.common.infra;
 
+import kitchenpos.common.infra.PurgomalumClient;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
