+++ src/main/java/kitchenpos/order/infra/DefaultKitchenridersClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.order.infra;
 
 import org.springframework.stereotype.Component;
